python-chatdownloader/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "chatdownloader"
version = "0.1.0"
description = "Downloads chat from the latest VOD"
authors = ["vanorsigma"]
packages = [{include="*", from="src"}]

[tool.poetry.dependencies]
python = "^3.10.12"
requests = "^2.31.0"
dataclasses-json = "0.6.5"
twitchapi = "^4.2.0"
pygithub = "^2.3.0"
numpy = "^2.0.0"

[tool.poetry.dev-dependencies]
pylint = "*"
black = "^24.0.0"
flake8 = "*"
